Palette: move source file to top-level
